+++ src/components/header/ExtendedNav/ExtendedNav.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { render } from '@testing-library/react'
+import { render, fireEvent } from '@testing-library/react'
 
 import { ExtendedNav } from './ExtendedNav'
 
@@ -21,10 +21,15 @@ const testSecondaryItems = [
   </a>,
 ]
 
-describe('Title component', () => {
+const onToggleMobileNav = (): void => {
+  /* mock submit fn */
+}
+
+describe('ExtendedNav component', () => {
   it('renders without errors', () => {
     const { queryByRole } = render(
       <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
         primaryItems={testPrimaryItems}
         secondaryItems={testSecondaryItems}
       />
@@ -35,6 +40,7 @@ describe('Title component', () => {
   it('renders primary items', () => {
     const { getByText } = render(
       <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
         primaryItems={testPrimaryItems}
         secondaryItems={testSecondaryItems}
       />
@@ -46,6 +52,7 @@ describe('Title component', () => {
   it('renders secondary items', () => {
     const { getByText } = render(
       <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
         primaryItems={testPrimaryItems}
         secondaryItems={testSecondaryItems}
       />
@@ -57,10 +64,49 @@ describe('Title component', () => {
   it('renders nav button', () => {
     const { getByTestId } = render(
       <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
         primaryItems={testPrimaryItems}
         secondaryItems={testSecondaryItems}
       />
     )
     expect(getByTestId('navButton')).toBeInTheDocument()
+  })
+
+  it('implements an onClick handler', () => {
+    const onToggleMobileNav = jest.fn()
+    const { getByTestId } = render(
+      <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
+        primaryItems={testPrimaryItems}
+        secondaryItems={testSecondaryItems}
+      />
+    )
+
+    fireEvent.click(getByTestId('navButton'))
+    expect(onToggleMobileNav).toHaveBeenCalledTimes(1)
+  })
+
+  it('renders the is-visible class when mobileExpanded is true', () => {
+    const { container } = render(
+      <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
+        primaryItems={testPrimaryItems}
+        secondaryItems={testSecondaryItems}
+        mobileExpanded={true}
+      />
+    )
+    expect(container.querySelector('.is-visible')).toBeInTheDocument()
+  })
+
+  it('does not render the is-visible class when mobileExpanded is false', () => {
+    const { container } = render(
+      <ExtendedNav
+        onToggleMobileNav={onToggleMobileNav}
+        primaryItems={testPrimaryItems}
+        secondaryItems={testSecondaryItems}
+        mobileExpanded={false}
+      />
+    )
+    expect(container.querySelector('.is-visible')).toEqual(null)
   })
 })