+++ src/components/header/Header/Header.tsx
@@ -2,17 +2,23 @@ import React from 'react'
 import classnames from 'classnames'
 
 interface HeaderProps {
+  basic?: boolean
+  extended?: boolean
   children: React.ReactNode
 }
 
 export const Header = (
   props: HeaderProps & React.HtmlHTMLAttributes<HTMLElement>
 ): React.ReactElement => {
-  const { children, ...headerProps } = props
+  const { basic, extended, children, className, ...headerProps } = props
 
   const classes = classnames(
-    'usa-header usa-header--basic',
-    headerProps.className
+    'usa-header',
+    {
+      'usa-header--basic': basic,
+      'usa-header--extended': extended,
+    },
+    className
   )
 
   return (
@@ -20,4 +26,6 @@ export const Header = (
       {children}
     </header>
   )
+
+  //TODO Add tests for basic and extended classes
 }