+++ src/components/header/MegaMenu/MegaMenu.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { List } from '../List/List'
+import { NavList } from '../NavList/NavList'
 
 type MegaMenuProps = {
   items: React.ReactNode[][]
@@ -10,11 +10,14 @@ type MegaMenuProps = {
 export const MegaMenu = (props: MegaMenuProps): React.ReactElement => {
   const { items, isOpen } = props
   return (
-    <div className="usa-nav__submenu usa-megamenu" hidden={!isOpen}>
+    <div
+      className="usa-nav__submenu usa-megamenu"
+      hidden={!isOpen}
+      data-testid="megamenu">
       <div className="grid-row grid-gap-4">
         {items.map((listItems, i) => (
           <div className="usa-col" key={`subnav_col_${i}`}>
-            <List
+            <NavList
               items={listItems}
               keyPrefix="subnav_item"
               ulClass="usa-nav__submenu-list"