+++ src/components/header/NavDropDown/NavDropDown.tsx
@@ -12,6 +12,7 @@ export const NavDropDown = (props: NavDropDownProps): React.ReactElement => {
   const { label, id, isOpen, onToggle } = props
   return (
     <NavButton
+      data-testid="navButton"
       className="usa-accordion__button usa-nav__link"
       aria-expanded={isOpen}
       aria-controls={id}